# Implement `check_bool()` in C