--- a/R/rlang-package.R
+++ b/R/rlang-package.R
@@ -9,3 +9,6 @@ on_load({
 compiled_by_gcc <- function() {
   .Call(ffi_compiled_by_gcc)
 }
+
+#' @rawNamespace export(ffi_standalone_check_is_bool)
+NULL