POC for implementing a function in a standalone file in C.

- The FFI symbol is exported and we'll commit to supporting it for a full life cycle. If deprecated, it will go through a normal deprecation process using the lifecycle package.

- If a new version is needed for an interface change, we'll export a new versioned FFI symbol.